Delay printing descriptions
Due to the unsorted nature of dicts, it is not the most aestetic to have
the description printed before more important information.

Ensure the description is printed last.

N.B. This is an intermediate step towards completely customisable
printing of available data.
The next step would be to allow all fields to be arrangeable, taking
type into consideration.